feat: 라인 API CRUD 테스트 진행중

+++ src/main/java/com/huh/BaekJoonSupporter/line/LineRepository.java
@@ -2,6 +2,6 @@ package com.huh.BaekJoonSupporter.line;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LineRepository extends JpaRepository<Line, Integer> {
+public interface LineRepository extends JpaRepository<Line, Long> {
 
 }
